Fixed button read error, readjusted code
Done at school.

All testing was done on a spare cortex and was not tested on the actual
robot. Testing led to discovery of why buttons were not read (case
error) which has solved that.

The next fix should be to figure out why the program won't run fully,
perhaps changing control from a task to a function that iterates a few
times and then ends so that the main task doesn't get held up.